requirements.txt
requests==2.28.1
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
schedule==1.1.0